+++ src/compas_rrc/__init__.py
@@ -138,7 +138,7 @@ from compas_rrc.watch import (
     StopWatch
 )
 
-
+__all_plugins__ = ['compas_rrc.__install']
 __all__ = [
     '__url__',
     '__version__',