+++ src/compas_rrc/__install.py
@@ -1,0 +1,6 @@
+import compas.plugins
+
+
+@compas.plugins.plugin(category='install')
+def installable_rhino_packages():
+    return ['compas_rrc']