+++ CHANGELOG.rst
@@ -5,6 +5,18 @@ Changelog
 Unreleased
 ----------
 
+**Added**
+
+* Prepared github actions for continuous integration
+* Added compas plugin for automatic Rhino install
+
+**Changed**
+
+**Fixed**
+
+**Deprecated**
+
+**Removed**
 
 1.0.0
 -------